refactor: all double values to float

+++ app/src/main/java/com/example/daggerhilttest/models/v1/HourlyForecastLocal.kt
@@ -1,7 +1,7 @@
 package com.example.daggerhilttest.models.v1
 
 data class HourlyForecastLocal(
-    val temp:Double? = null,
+    val temp: Float? = null,
     val iconUrl:String? = null,
     val timeString:String? = null,
     val timeFloatHrs:Float? = null
