+++ app/src/main/java/com/example/daggerhilttest/models/v2/CurrentWeatherDataV2.kt
@@ -5,8 +5,8 @@ import com.example.daggerhilttest.models.v1.WeatherExtraDetailItem
 data class CurrentWeatherDataV2(
     val dateTime: String,
     val dateTimeEpoch: String,
-    val temperature: Double,
-    val feelsLikeTemp: Double,
+    val temperature: Float,
+    val feelsLikeTemp: Float,
     val weatherCondition: String,
     val weatherDetailItems: List<WeatherExtraDetailItem>,
     val icon: String,
